requirements.txt

appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coverage==5.5
cycler==0.10.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.10.0
django-heroku==0.3.1
django-silk==4.1.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
filelock==3.0.12
gprof2dot==2021.2.21
gunicorn==20.1.0
idna==2.10
isort==5.8.0
Jinja2==2.11.3
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1

psycopg2==2.8.6
pycodestyle==2.7.0
Pygments==2.8.1
PyJWT==1.7.1
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pymongo==3.12.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
schedule==1.1.0
sqlparse==0.2.4
toml==0.10.2
twilio==6.58.0
urllib3==1.26.4
virtualenv==20.4.3
whitenoise==6.0.0
wrapt==1.12.1
